api/requirements.txt
# api/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==4.6.0
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
authlib==1.2.1
httpx==0.25.2

# Pydantic & Validation
pydantic==2.5.0
pydantic-settings==2.0.3
email-validator==2.1.0

# Background Jobs
celery[redis]==5.3.4
flower==2.0.1

# File Upload & Storage
python-multipart==0.0.6
boto3==1.34.0
pillow==10.1.0

# Notifications
firebase-admin==6.2.0
sendgrid==6.10.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
pendulum==2.1.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
factory-boy==3.3.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1